I fixed the compilation errors in `lib/core/gamification/gamification_engine.dart` and `lib/core/sharing/share_service.dart` by importing `MinqLogger` and using its static methods for logging. After these fixes, I ran `flutter test` and all 16 tests passed.

+++ lib/presentation/screens/stats_screen.dart
@@ -558,14 +558,14 @@ Widget _buildWeeklyProgressCard(
     ),
     _RingMetric(
       label: '合計時間',
-      value: totalMinutes / 60,
+      value: (totalMinutes / 60).toDouble(),
       unit: '時間',
       progress:
           ((totalMinutes / 60) / (weeklyGoal * dailyGoalMinutes / 60)).toDouble(),
     ),
     _RingMetric(
       label: '平均時間',
-      value: averageMinutes,
+      value: averageMinutes.toDouble(),
       unit: '分',
       progress: (averageMinutes / dailyGoalMinutes).toDouble(),
     ),
